feat: Adds a theme provider and a main theme, refactors the code to modularize components
