+++ .env
@@ -1,2 +1,2 @@
-REACT_APP_BACKEND_SERVER=http://ec2-3-133-82-70.us-east-2.compute.amazonaws.com:4000/graphql
+REACT_APP_BACKEND_SERVER=http://ec2-18-222-254-122.us-east-2.compute.amazonaws.com:4000/graphql
 LOCAL_BACKEND_SERVER=http://localhost:4000/graphql
