--- src/components/Chat/ChatStyle.js
@@ -1,53 +1,0 @@
-export default {
-  backDrop: {
-    backgroundColor: "#8ec4f5",
-  },
-  root: {
-    textAlign: "center",
-    backgroundColor: "#d9eafa",
-    minHeight: "100vh",
-    display: "flex",
-    flexDirection: "column",
-    marginLeft: "20vw",
-    marginRight: "20vw",
-    paddingRight: "15px",
-    paddingLeft: "15px",
-    color: "#333",
-  },
-  header: {
-    maginTop: "10px",
-    alignItems: "center",
-    justifyContent: "space-between",
-    display: "flex",
-  },
-  messageText: {
-    textAlign: "left",
-    fontSize: "medium",
-    marginLeft: "10px",
-    whiteSpace: "pre-wrap",
-    overflowWrap: "break-word",
-  },
-  messageTitle: {
-    display: "flex",
-    alignItems: "flex-end",
-  },
-  creatorName: {
-    textAlign: "left",
-    fontSize: "large",
-    fontWeight: "bold",
-  },
-  timeStamp: {
-    margin: "2px 4px",
-    fontSize: "10px",
-  },
-  messageBlock: {
-    width: "100%",
-  },
-  messagesBlock: {
-    maxHeight: "Calc(100vh - 310px)",
-    overflow: "auto",
-  },
-  submitButton: {
-    backgroundColor: "#82b7e8",
-  },
-};
