--- src/components/Login/LoginStyle.js
@@ -1,25 +1,0 @@
-export default {
-  backDrop: {
-    backgroundColor: "#8ec4f5",
-  },
-  root: {
-    textAlign: "center",
-    backgroundColor: "#d9eafa",
-    minHeight: "100vh",
-    display: "flex",
-    flexDirection: "column",
-    marginLeft: "20vw",
-    marginRight: "20vw",
-    paddingRight: "15px",
-    paddingLeft: "15px",
-    color: "#333",
-  },
-  header: {
-    maginTop: "10px",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  submitButton: {
-    backgroundColor: "#8ec4f5",
-  },
-};
