--- src/components/UserHeader/UserHeaderStyle.js
@@ -1,12 +1,0 @@
-export default {
-  loginButton: {
-    backgroundColor: "#82b7e8",
-  },
-  username: {
-    margin: "8px",
-    fontWeight: "700",
-  },
-  userRow: {
-    display: "flex",
-  },
-};
